kyd_downloader/pyproject.toml
[tool.poetry]
name = "kyd-downloader"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kyd_downloader"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
gunicorn = "^20.1.0"
lxml = "^4.9.2"
bizdays = "^1.0.8"
google-cloud-storage = "^2.10.0"
google-cloud-datastore = "^2.16.0"
cssselect = "^1.2.0"
pytz = "^2023.3"
pandas = "^2.0.2"
kyd = {git = "https://github.com/datalab42/kyd.git"}

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.15.3"
matplotlib = "^3.7.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
